feat(core): pass `refresh` to all events
