+++ packages/autocomplete-core/src/types/AutocompleteApi.ts
@@ -1,0 +1,20 @@
+import { AutocompletePropGetters } from './AutocompletePropGetters';
+import { AutocompleteSetters } from './AutocompleteSetters';
+
+export type BaseItem = Record<string, unknown>;
+
+export interface AutocompleteScopeApi<TItem extends BaseItem>
+  extends AutocompleteSetters<TItem> {
+  /**
+   * Triggers a search to refresh the state.
+   */
+  refresh: () => Promise<void>;
+}
+
+export type AutocompleteApi<
+  TItem extends BaseItem,
+  TEvent = Event,
+  TMouseEvent = MouseEvent,
+  TKeyboardEvent = KeyboardEvent
+> = AutocompleteScopeApi<TItem> &
+  AutocompletePropGetters<TItem, TEvent, TMouseEvent, TKeyboardEvent>;
