+++ packages/autocomplete-core/src/types/AutocompleteCollection.ts
@@ -1,0 +1,12 @@
+import { BaseItem } from './AutocompleteApi';
+import { InternalAutocompleteSource } from './AutocompleteSource';
+
+export interface AutocompleteCollection<TItem extends BaseItem> {
+  source: InternalAutocompleteSource<TItem>;
+  items: TItem[];
+}
+
+export interface AutocompleteCollectionItemsArray<TItem extends BaseItem> {
+  source: InternalAutocompleteSource<TItem>;
+  items: TItem[][];
+}
