+++ packages/autocomplete-core/src/types/AutocompletePlugin.ts
@@ -1,10 +1,7 @@
-import {
-  AutocompleteOptions,
-  OnHighlightParams,
-  OnSelectParams,
-  BaseItem,
-} from './api';
-import { AutocompleteSetters } from './setters';
+import { BaseItem } from './AutocompleteApi';
+import { AutocompleteOptions } from './AutocompleteOptions';
+import { AutocompleteSetters } from './AutocompleteSetters';
+import { OnSelectParams, OnHighlightParams } from './AutocompleteSource';
 
 type PluginSubscriber<TParams> = (params: TParams) => void;
