--- packages/autocomplete-core/src/types/api.ts
@@ -1,295 +1,0 @@
-import { MaybePromise } from '@algolia/autocomplete-shared';
-
-import { AutocompleteAccessibilityGetters } from './getters';
-import { AutocompletePlugin } from './plugins';
-import { AutocompleteSetters } from './setters';
-import { AutocompleteState } from './state';
-
-export type BaseItem = Record<string, unknown>;
-
-export interface AutocompleteApi<
-  TItem extends BaseItem,
-  TEvent = Event,
-  TMouseEvent = MouseEvent,
-  TKeyboardEvent = KeyboardEvent
-> extends AutocompleteSetters<TItem>,
-    AutocompleteAccessibilityGetters<
-      TItem,
-      TEvent,
-      TMouseEvent,
-      TKeyboardEvent
-    > {
-  /**
-   * Triggers a search to refresh the state.
-   */
-  refresh: AutocompleteRefresh;
-}
-
-export type AutocompleteRefresh = () => Promise<void>;
-
-export interface AutocompleteCollection<TItem extends BaseItem> {
-  source: InternalAutocompleteSource<TItem>;
-  items: TItem[];
-}
-export interface AutocompleteCollectionItemsArray<TItem extends BaseItem> {
-  source: InternalAutocompleteSource<TItem>;
-  items: TItem[][];
-}
-
-export interface GetSourcesParams<TItem extends BaseItem>
-  extends AutocompleteSetters<TItem> {
-  query: string;
-  state: AutocompleteState<TItem>;
-  refresh: AutocompleteRefresh;
-}
-
-interface ItemParams<TItem extends BaseItem> {
-  item: TItem;
-  itemInputValue: ReturnType<
-    InternalAutocompleteSource<TItem>['getItemInputValue']
-  >;
-  itemUrl: ReturnType<InternalAutocompleteSource<TItem>['getItemUrl']>;
-  source: InternalAutocompleteSource<TItem>;
-}
-
-export interface OnSelectParams<TItem extends BaseItem>
-  extends ItemParams<TItem>,
-    AutocompleteSetters<TItem> {
-  state: AutocompleteState<TItem>;
-  event: any;
-}
-
-export type OnHighlightParams<TItem extends BaseItem> = OnSelectParams<TItem>;
-
-interface OnSubmitParams<TItem extends BaseItem>
-  extends AutocompleteSetters<TItem> {
-  state: AutocompleteState<TItem>;
-  event: any;
-}
-
-type OnResetParams<TItem extends BaseItem> = OnSubmitParams<TItem>;
-
-interface OnInputParams<TItem extends BaseItem>
-  extends AutocompleteSetters<TItem> {
-  query: string;
-  state: AutocompleteState<TItem>;
-  refresh: AutocompleteRefresh;
-}
-
-export interface AutocompleteSource<TItem extends BaseItem> {
-  // This allows flavors to pass other keys to their source.
-  // Example: `templates` in the JavaScript API
-  // [key: string]: unknown;
-  /**
-   * Get the string value of the item. The value is used to fill the search box.
-   */
-  getItemInputValue?({
-    item,
-    state,
-  }: {
-    item: TItem;
-    state: AutocompleteState<TItem>;
-  }): string;
-  /**
-   * Get the URL of a item. The value is used to create default navigation features for
-   * `onClick` and `onKeyDown`.
-   */
-  getItemUrl?({
-    item,
-    state,
-  }: {
-    item: TItem;
-    state: AutocompleteState<TItem>;
-  }): string | undefined;
-  /**
-   * Function called when the input changes. You can use this function to filter/search the items based on the query.
-   */
-  getItems(params: GetSourcesParams<TItem>): MaybePromise<TItem[] | TItem[][]>;
-  /**
-   * Function called when an item is selected.
-   */
-  onSelect?(params: OnSelectParams<TItem>): void;
-  /**
-   * Function called when an item is highlighted.
-   *
-   * An item is highlighted either via keyboard navigation or via mouse over.
-   * You can trigger different behaviors based on the event `type`.
-   */
-  onHighlight?(params: OnHighlightParams<TItem>): void;
-}
-
-export type InternalAutocompleteSource<TItem extends BaseItem> = {
-  [KParam in keyof AutocompleteSource<TItem>]-?: AutocompleteSource<TItem>[KParam];
-};
-
-export type GetSources<TItem extends BaseItem> = (
-  params: GetSourcesParams<TItem>
-) => Promise<Array<InternalAutocompleteSource<TItem>>>;
-
-export type Environment =
-  | Window
-  | {
-      [prop: string]: unknown;
-      addEventListener: Window['addEventListener'];
-      removeEventListener: Window['removeEventListener'];
-      setTimeout: Window['setTimeout'];
-      clearTimeout: Window['clearTimeout'];
-      document: Window['document'];
-      location: {
-        assign: Location['assign'];
-      };
-      open: Window['open'];
-    };
-
-interface Navigator<TItem extends BaseItem> {
-  /**
-   * Called when a URL should be open in the current page.
-   */
-  navigate(params: {
-    itemUrl: string;
-    item: TItem;
-    state: AutocompleteState<TItem>;
-  }): void;
-  /**
-   * Called when a URL should be open in a new tab.
-   */
-  navigateNewTab(params: {
-    itemUrl: string;
-    item: TItem;
-    state: AutocompleteState<TItem>;
-  }): void;
-  /**
-   * Called when a URL should be open in a new window.
-   */
-  navigateNewWindow(params: {
-    itemUrl: string;
-    item: TItem;
-    state: AutocompleteState<TItem>;
-  }): void;
-}
-
-export interface AutocompleteOptions<TItem extends BaseItem> {
-  /**
-   * Whether to consider the experience in debug mode.
-   *
-   * The debug mode is useful when developing because it doesn't close
-   * the panel when the blur event occurs.
-   *
-   * @default false
-   */
-  debug?: boolean;
-  /**
-   * The unique Autocomplete ID to create accessible attributes.
-   *
-   * It is incremented by default when creating a new Autocomplete instance.
-   *
-   * @default "autocomplete-0"
-   */
-  id?: string;
-  /**
-   * Function called when the internal state changes.
-   */
-  onStateChange?(props: {
-    state: AutocompleteState<TItem>;
-    prevState: AutocompleteState<TItem>;
-  }): void;
-  /**
-   * The text that appears in the search box input when there is no query.
-   */
-  placeholder?: string;
-  /**
-   * Whether to focus the search box when the page is loaded.
-   *
-   * @default false
-   */
-  autoFocus?: boolean;
-  /**
-   * The default item index to pre-select.
-   *
-   * @default null
-   */
-  defaultSelectedItemId?: number | null;
-  /**
-   * Whether to open the panel on focus when there's no query.
-   *
-   * @default false
-   */
-  openOnFocus?: boolean;
-  /**
-   * The number of milliseconds that must elapse before the autocomplete
-   * experience is stalled.
-   *
-   * @default 300
-   */
-  stallThreshold?: number;
-  /**
-   * The initial state to apply when autocomplete is created.
-   */
-  initialState?: Partial<AutocompleteState<TItem>>;
-  /**
-   * The sources to get the suggestions from.
-   */
-  getSources?(
-    params: GetSourcesParams<TItem>
-  ): MaybePromise<Array<AutocompleteSource<TItem>>>;
-  /**
-   * The environment from where your JavaScript is running.
-   * Useful if you're using autocomplete in a different context than
-   * `window`.
-   *
-   * @default window
-   */
-  environment?: Environment;
-  /**
-   * Navigator API to redirect the user when a link should be opened.
-   */
-  navigator?: Partial<Navigator<TItem>>;
-  /**
-   * The function called to determine whether the panel should open.
-   */
-  shouldPanelShow?(params: { state: AutocompleteState<TItem> }): boolean;
-  /**
-   * The function called when the Autocomplete form is submitted.
-   */
-  onSubmit?(params: OnSubmitParams<TItem>): void;
-  /**
-   * The function called when the Autocomplete form is reset.
-   */
-  onReset?(params: OnResetParams<TItem>): void;
-  /**
-   * The function called when the input changes.
-   *
-   * This turns the experience in controlled mode, leaving you in charge of
-   * updating the state.
-   */
-  onInput?(params: OnInputParams<TItem>): void;
-  /**
-   * The array of plugins.
-   */
-  plugins?: Array<AutocompletePlugin<TItem, unknown>>;
-}
-
-// Props manipulated internally with default values.
-export interface InternalAutocompleteOptions<TItem extends BaseItem>
-  extends AutocompleteOptions<TItem> {
-  debug: boolean;
-  id: string;
-  onStateChange(props: {
-    state: AutocompleteState<TItem>;
-    prevState: AutocompleteState<TItem>;
-  }): void;
-  placeholder: string;
-  autoFocus: boolean;
-  defaultSelectedItemId: number | null;
-  openOnFocus: boolean;
-  stallThreshold: number;
-  initialState: AutocompleteState<TItem>;
-  getSources: GetSources<TItem>;
-  environment: Environment;
-  navigator: Navigator<TItem>;
-  plugins: Array<AutocompletePlugin<TItem, unknown>>;
-  shouldPanelShow(params: { state: AutocompleteState<TItem> }): boolean;
-  onSubmit(params: OnSubmitParams<TItem>): void;
-  onReset(params: OnResetParams<TItem>): void;
-  onInput?(params: OnInputParams<TItem>): void | Promise<any>;
-}
