+++ packages/autocomplete-core/src/types/index.ts
@@ -1,7 +1,11 @@
-export * from './api';
-export * from './getters';
-export * from './plugins';
-export * from './setters';
-export * from './state';
-export * from './store';
-export * from './subscribers';
+export * from './AutocompleteApi';
+export * from './AutocompleteCollection';
+export * from './AutocompleteEnvironment';
+export * from './AutocompleteOptions';
+export * from './AutocompleteSource';
+export * from './AutocompletePropGetters';
+export * from './AutocompletePlugin';
+export * from './AutocompleteSetters';
+export * from './AutocompleteState';
+export * from './AutocompleteStore';
+export * from './AutocompleteSubscribers';
