--- packages/autocomplete-core/src/types/subscribers.ts
@@ -1,10 +1,0 @@
-import { BaseItem, OnHighlightParams, OnSelectParams } from './api';
-
-export type Subscriber<TItem extends BaseItem> = {
-  onSelect(params: OnSelectParams<TItem>): void;
-  onHighlight(params: OnHighlightParams<TItem>): void;
-};
-
-export type Subscribers<TItem extends BaseItem> = Array<
-  Partial<Subscriber<TItem>>
->;
